% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{add_days}
\alias{add_days}
\title{Adds a number of days to a date}
\usage{
add_days(year, month, day, bce, no_days)
}
\arguments{
\item{year}{A positive integer}

\item{month}{A positive integer}

\item{day}{A positive integer}

\item{bce}{Logical variable, indicates if the date is Before Common Era}

\item{no_days}{A positive integer. The number of days to count from the
date passed}
}
\description{
It adds a number of days to the passed date, and returns a new date. It
performs the calculations independently from any R date/time function.
An advantage of this function is that it is accepts dates older than year 
1 CE.  It uses calendar and not astronomical year numbering.  It only
adds days, passing a negative no_days returns an error.
}
\examples{

add_days(3114, 8, 11, TRUE, 1872000)

# These should return the same date
as.Date("2000-03-03") + 1000
add_days(2000, 3, 3, FALSE, 1000)

}
