% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAYA_cell_type.R
\name{annotate_cells}
\alias{annotate_cells}
\title{Annotate individual cells}
\usage{
annotate_cells(annotation, cluster_annot)
}
\arguments{
\item{Annotation}{Vector, containing cluster annotation for individual cells}

\item{clusters_annot}{Dataframe giving corresponding annotation for each cluster}
}
\value{
Vector containing annotation for individual cells
}
\description{
Annotate individual cells
}
