% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{average_by_cluster}
\alias{average_by_cluster}
\title{Average matrix by column}
\usage{
average_by_cluster(mat, annotation)
}
\arguments{
\item{mat}{Numeric/sparse matrix (cells as columns)}

\item{annotation}{Vector containg cell annotation}
}
\value{
Numeric/sparse matrix.
}
\description{
Compute for each row the average value by provided annotation
}
