% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{build_activity_mat}
\alias{build_activity_mat}
\title{Extract activity matrix from PCA object}
\usage{
build_activity_mat(PCA_object, scaled = T)
}
\arguments{
\item{PCA_object}{List, output from run_activity_analysis().}

\item{scaled}{Boolean. By default, using scores scaled between 0 and 1. Set to FALSE if you want to keep raw projections of cells on PCs.}
}
\value{
Numeric matrix, with the different activation ways as rows and cells as columns and containing the activity score.
}
\description{
Extract activity matrix from PCA object
}
