% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cal_z_score}
\alias{cal_z_score}
\title{Scale vector between 0 and 1}
\usage{
cal_z_score(x)
}
\arguments{
\item{X}{Numeric vector.}
}
\value{
Numeric vector.
}
\description{
Changes the range of values of a vector keeping the shape of the distribution. Minimum value is set to 0 and maximum to 1.
}
\examples{
vect <- rnorm(n = 50, mean = 25, sd = 3)
cal_z_score(vect)
}
