% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter.R
\name{diff_days}
\alias{diff_days}
\title{Number of days between two gregorian dates}
\usage{
diff_days(year_1, month_1, day_1, bce_1, year_2, month_2, day_2, bce_2)
}
\arguments{
\item{year_1}{A positive integer}

\item{month_1}{A positive integer}

\item{day_1}{A positive integer}

\item{bce_1}{Logical variable, indicates if the date is Before Common Era}

\item{year_2}{A positive integer}

\item{month_2}{A positive integer}

\item{day_2}{A positive integer}

\item{bce_2}{Logical variable, indicates if the date is Before Common Era}
}
\description{
It determines the number of days between two gregorian dates.  It works 
independently from any R date/time function. An advantage of this function 
is that it is accepts dates older than year 1 CE.  It uses calendar and 
not astronomical year numbering.
}
\examples{

diff_days(3114, 8, 11, TRUE, 2012, 12, 21, FALSE)

}
