% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informativity.R
\name{.activity_assignmentThreshold}
\alias{.activity_assignmentThreshold}
\title{Auxiliary function to determine if a vector is informative.}
\usage{
.activity_assignmentThreshold(activity)
}
\arguments{
\item{activity}{Numeric vector, corresponding to activity scores across cells, containing values between 0 and 1.}
}
\value{
Double, threshold distinguishing cell populations.
}
\description{
Informativity is determined by our ability to set a threshold, a value comprised between the minimum and maximum value of the distribution and that
separates two distinct groups of observations: the ones above threshold and the ones under. If the threshold is above 1, cells couldn't be distinguished
based on activity, the module is uninformative.
}
