% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAYA_cell_type.R
\name{generate_cell_type_annotation}
\alias{generate_cell_type_annotation}
\title{Generate cell type annotation for each cell and provide complementary results to interpret results}
\usage{
generate_cell_type_annotation(PCA_obj, thr = 0, compute_umap = T)
}
\arguments{
\item{PCA_obj}{Vector, containing cluster annotation for individual cells}

\item{thr}{Minimum average activity by cluster required to be assigned an annotation. Default 0, increase to be more stringent on assignation confidence}
}
\value{
List containing cell type annotation, matrix of average score by Leiden cluster and cell type used for attribution, cell annotation with Leiden clusters, activity matrix and UMAP computed on activity matrix.
}
\description{
Generate cell type annotation for each cell and provide complementary results to interpret results
}
