% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_palette}
\alias{generate_palette}
\title{Generate custom color palette}
\usage{
generate_palette(length)
}
\arguments{
\item{length}{Number of colors required}
}
\value{
Vector of colors
}
\description{
Generate custom color palette
}
