% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_exploration.R
\name{generate_summary_specific_modules}
\alias{generate_summary_specific_modules}
\title{Generate summary of specific pathways and top contributing genes given an annotation}
\usage{
generate_summary_specific_modules(
  PCA_object,
  file = NULL,
  meta,
  annot_name,
  top_genes = 10,
  min_contrib = 0.1
)
}
\arguments{
\item{PCA_object}{List, output from run_activity_analysis().}

\item{file}{Character, path to file if you want the analysis stored as xlsx file with one sheet by cluster.}

\item{meta}{Dataframe containing metadata, with cell labels as rownames.}

\item{annot_name}{Character, name of the column in meta that should be used as annotation for cells.}

\item{top_genes}{Integer, number of top genes contributing to PC to print.}

\item{min_contrib}{Double, minimum contribution/weight of a gene to the PC. It overrides top_genes parameter and only top_genes contributing above min_contrib will be printed.}

\item{noise}{Double, value between 0 and 1 indicating the maximum specificity you allow a non-specific cluster to have. Can be increased if no pathways appear as specific.}
}
\value{
List, one assay by cluster defined by annotation, and each displaying a table with the pathway, the top contributing genes and the pathway specificity.
}
\description{
Generate summary of specific pathways and top contributing genes given an annotation
}
