% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{gregorian_to_mayan}
\alias{gregorian_to_mayan}
\title{Converts a gregorian date to mayan long count date}
\usage{
gregorian_to_mayan(date = NULL, input = c("calendar", "named_vector"),
  output = c("character", "named_vector"))
}
\arguments{
\item{date}{A character vector}

\item{input}{A character vector}

\item{output}{A character vector}
}
\description{
Converts a gregorian date to mayan long count date
}
\examples{

gregorian_to_mayan("August 11, 3114 BCE")
gregorian_to_mayan("May 19, 143, CE")

}
