% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{gregorian_to_mayan2}
\alias{gregorian_to_mayan2}
\title{Converts a gregorian date to mayan long count date}
\usage{
gregorian_to_mayan2(year, month, day, bce)
}
\arguments{
\item{year}{A positive integer}

\item{month}{A positive integer}

\item{day}{A positive integer}

\item{bce}{Logical variable, indicates if the date is Before Common Era}
}
\description{
It converts a gregorian calendar date into mayan long count date. It expects
4 arguments: year number, month number, day number, and if the date is Before 
Common Era (BCE). It expects a calendar year not as an astronomical year.  
Dates before year 1, have to be passed by setting bce to TRUE.
 
It returns a vector with 5 named values. The lowest count is kin, which is equivalent 
to a day. 20 days, or kins is a winal. 18 winals is a tun. 20 tuns is a katun. 
20 katuns is a baktun.  This means that a baktun represents a cycle of 144,000 days 
(https://en.wikipedia.org/wiki/Maya_calendar).
}
\details{
The count begins from what has been determined to be the equivalent of mayan long count
date 0.0.0.0.0.  The equivalent in the Gregorian calendar is August 11 3114 BCE.
}
\examples{

# August 11 3114 BCE
gregorian_to_mayan2(3114, 8, 11, TRUE)
# May 19 143 CE
gregorian_to_mayan2(143, 5, 19, FALSE)

}
