% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_leap_year}
\alias{is_leap_year}
\title{Determines if it is a leap year}
\usage{
is_leap_year(year)
}
\arguments{
\item{year}{The year to evaluate. It expects a integer number. It
also expects the astronomical number of years older than 1. For example,
to evaluate year 401 BCE, use -400.}
}
\description{
It follows the established rules to figure out if is a leap year 
(https://en.wikipedia.org/wiki/Leap_year).  It uses astronomial
year numbering for years before the Gregorian calendar was established.  
(https://en.wikipedia.org/wiki/Astronomical_year_numbering)
}
\details{
The year is designated as leap if: the year number is divisible by 400,
or if the year number is divisible by 4, but it is not divisible by
100.
}
\examples{

is_leap_year(2012)
is_leap_year(-401)
# To evaluate multiple years
as.integer(lapply(2000:2005, is_leap_year))

}
