% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{jaccard}
\alias{jaccard}
\title{Compute Jaccard distance from kNN matrix}
\usage{
jaccard(m)
}
\arguments{
\item{m}{Numeric/sparse matrix.}
}
\value{
Numeric/sparse matrix.
}
\description{
Compute Jaccard distance from kNN matrix
}
