% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{knn_jaccard}
\alias{knn_jaccard}
\title{Compute Jaccard distance from kNN results}
\usage{
knn_jaccard(knn)
}
\arguments{
\item{knn}{Numeric matrix (cells as rows, index of neighbors as columns (20 columns if k=20))}
}
\value{
Numeric/sparse matrix.
}
\description{
Compute Jaccard distance from kNN results
}
