% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logcpmNormalization}
\alias{logcpmNormalization}
\title{Transform raw count matrix into logCPM matrix}
\usage{
logcpmNormalization(X)
}
\arguments{
\item{X}{Numeric matrix.}
}
\value{
Numeric matrix.
}
\description{
Converts continuous BMI variable to BMI groups.
}
\examples{
raw_mat <- matrix(1:20, ncol=5, nrow=4)
norm_mat<-logcpmNormalization(raw_mat)

}
