% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAYA_pathways.R
\name{MAYA_pathway_analysis}
\alias{MAYA_pathway_analysis}
\title{Run MAYA for unsupervised pathway acitivity analysis}
\usage{
MAYA_pathway_analysis(
  expr_mat,
  modules_list = NULL,
  min_cells_pct = 0.05,
  is_logcpm = T,
  nCores = 1,
  min_genes = 10,
  max_contrib = 0.5,
  compute_umap = T
)
}
\arguments{
\item{expr_mat}{Raw count matrix. If normalized, set is_logcpm to T}

\item{modules_list}{List with pathways associated with their genes. Can also set to "hallmark" or "kegg" to load corresponding MSigDB lists.}

\item{min_cells_pct}{Numeric between 0 and 1, minimum percentage of cells that should be above informativity threshold to consider activity score interesting.}

\item{is_logcpm}{Set to TRUE if data already normalized}

\item{nCores}{Number of cores to use. Set to 1 if working in a Windows environment, otherwise you can use the function detectCores() to find out how many cores are available.}

\item{max_contrib}{Numeric between 0 and 1, representing the maximum contribution to a PC allowed for a gene. Can influence how stringent mode selection is.}

\item{thr}{Minimum average activity by cluster required to be assigned an annotation. Default 0, increase to be more stringent on assignation confidence.}
}
\value{
List containing cell type annotation, matrix of average score by Leiden cluster and cell type used for attribution, cell annotation with MLeiden clusters, activity matrix and UMAp computed on activity matrix.
}
\description{
Run MAYA for unsupervised pathway acitivity analysis
}
