% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAYA_cell_type.R
\name{MAYA_predict_cell_types}
\alias{MAYA_predict_cell_types}
\title{Run MAYA for unsupervised cell annotation with cell type}
\usage{
MAYA_predict_cell_types(
  expr_mat,
  modules_list = NULL,
  min_cells_pct = 0.05,
  organs = NULL,
  is_logcpm = F,
  nCores = 1,
  thr = 0,
  max_contrib = 0.5,
  compute_umap = T,
  plot_heatmap = T,
  scale_before_pca = T
)
}
\arguments{
\item{expr_mat}{Raw count matrix. If normalized, set is_logcpm to T}

\item{modules_list}{List with cell types associated with their markers. If NULL, PanglaoDB will be loaded.}

\item{min_cells_pct}{Numeric between 0 and 1, minimum percentage of cells that should be above informativity threshold to consider activity score interesting.}

\item{organs}{Character to specify if list for a specific organ should be loaded. One of ("Pancreas","Brain","Lungs","Heart","Liver","Adrenal glands","GI tract","Reproductive","Kidney","Zygote","Thyroid","Embryo","Bone","Skin","Mammary gland","Eye","Olfactory system","Oral cavity","Thymus","Placenta","Urinary bladder"). "All" is another possible value, the 147 PanglaoDB cell types will be loaded.}

\item{is_logcpm}{Set to TRUE if data already normalized}

\item{nCores}{Number of cores to use. Set to 1 if working in a Windows environment, otherwise you can use the function detectCores() to find out how many cores are available.}

\item{thr}{Minimum average activity by cluster required to be assigned an annotation. Default 0, increase to be more stringent on assignation confidence.}

\item{max_contrib}{Numeric between 0 and 1, representing the maximum contribution to a mode allowed for a gene. Can influence how stringent mode selection is.}

\item{plot_heatmap}{Set to False to disable automatic display of activity matrix as a heatmap.}
}
\value{
List containing cell type annotation, matrix of average score by Leiden cluster and cell type used for attribution, cell annotation with Leiden clusters, activity matrix and UMAP computed on activity matrix.
}
\description{
Run MAYA for unsupervised cell annotation with cell type
}
