% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{mayan_to_gregorian}
\alias{mayan_to_gregorian}
\title{Converts mayan long count date to gregorian date}
\usage{
mayan_to_gregorian(baktun, katun, tun, winal, kin)
}
\arguments{
\item{baktun}{The number of 144,000 day cycles}

\item{katun}{The number of 7,200 day cycles}

\item{tun}{The number of 360 day cycles}

\item{winal}{The number of 20 day cycles}

\item{kin}{The number of days}
}
\description{
It converts a long count mayan calendar entry into a gregorian calendar date. It 
returns a single vector with 4 values: year number, month number, day number, 
and if the date is Before Common Era (BCE). The year is returned as a calendar year,
not as an astronomical year.  The year number will always be positive, the logical BCE 
variable indicates if the year is under year 1.
 
It expects 5 integer numbers representing each of the day cycles in the long count
calendar. The lowest count is kin, which is equivalent to a day. 20 days, or kins
is a winal. 18 winals is a tun. 20 tuns is a katun. 20 katuns is a baktun.  This means
that a baktun represents a cycle of 144,000 days.
}
\details{
The count starts at what has been determined to be the equivalent of mayan long count
date 0.0.0.0.0.  In the proleptic Gregorian calendar that is August 11 3114 BCE.
}
\examples{

mayan_to_gregorian(0, 0, 0, 0, 0)
mayan_to_gregorian(13, 0, 0, 0, 0)

}
