% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_exploration.R
\name{plot_heatmap_activity_mat}
\alias{plot_heatmap_activity_mat}
\title{Plots heatmap showing activity scores for each activation mode}
\usage{
plot_heatmap_activity_mat(
  activity_mat,
  meta = NULL,
  annot_name = NULL,
  cluster_cols = T,
  cluster_rows = T,
  clustering_distance = "euclidean",
  clustering_method = "ward.D2",
  fontsize = 5,
  colors_annot = NULL,
  fontsize_row = 5,
  fontsize_col = 5
)
}
\arguments{
\item{activity_mat}{Numeric matrix, containing activity for each activation mode of each pathway}

\item{meta}{Dataframe containing metadata, with cell labels as rownames.}

\item{annot_name}{Character, name of the column in meta that should be used to compute homogeneity for.}

\item{cluster_cols}{Boolean, set to TRUE to perform column clustering.}

\item{cluster_rows}{Boolean, set to TRUE to perform row clustering.}

\item{clustering_distance}{Character, distance to use to perform clustering, one of "euclidean","correlation".}

\item{clustering_method}{Character, one of "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC).}

\item{fontsize}{Integer, base fontsize for the plot.}

\item{colors_annot}{Vector of colors to use.}
}
\value{
heatmap plot
}
\description{
Plots heatmap showing activity scores for each activation mode
}
