% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_exploration.R
\name{plot_heatmap_pathway_genes}
\alias{plot_heatmap_pathway_genes}
\title{Plots heatmap showing individual gene expression of a pathway}
\usage{
plot_heatmap_pathway_genes(
  expr_mat,
  activity_mat = NULL,
  meta = NULL,
  annot_name = NULL,
  modules_list,
  module,
  cluster_cols = T,
  cluster_rows = T,
  clustering_distance = "euclidean",
  clustering_method = "ward.D2",
  fontsize = 5,
  colors_annot = NULL
)
}
\arguments{
\item{expr_mat}{Numeric matrix, containing counts, normalized or not.}

\item{activity_mat}{Numeric matrix, containing activity for each activation mode of each pathway.}

\item{meta}{Dataframe containing metadata, with cell labels as rownames.}

\item{annot_name}{Character, name of the column in meta that should be used to compute homogeneity for.}

\item{modules_list}{List, with module names as list names and character vectors containing genes in module. Output from read_gmt file for example.}

\item{module}{Character, pathway for which you want to plot expression.}

\item{cluster_cols}{Boolean, set to TRUE to perform column clustering.}

\item{cluster_rows}{Boolean, set to TRUE to perform row clustering.}

\item{clustering_distance}{Character, distance to use to perform clustering, one of "euclidean","correlation".}

\item{clustering_method}{Character, one of "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC).}

\item{fontsize}{Integer, base fontsize for the plot.}

\item{colors_annot}{Vector of colors to use.}
}
\value{
heatmap plot
}
\description{
Plots heatmap showing individual gene expression of a pathway
}
