% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_exploration.R
\name{plot_heatmap_pathway_top_contrib_genes}
\alias{plot_heatmap_pathway_top_contrib_genes}
\title{Plots heatmap of top contributing genes for each mode}
\usage{
plot_heatmap_pathway_top_contrib_genes(
  expr_mat,
  PCA_object,
  module,
  n = 10,
  meta = NULL,
  annot_name = NULL,
  cluster_cols = T,
  fontsize = 7,
  colors_annot = NULL
)
}
\arguments{
\item{expr_mat}{Numeric matrix, containing counts, normalized or not.}

\item{PCA_object}{List, output from run_activity_analysis().}

\item{module}{Character, module for which you want to plot activity}

\item{n}{Number of top contributors to plot for each mode}

\item{meta}{Dataframe containing metadata, with cell labels as rownames.}

\item{annot_name}{Character, name of the column in meta that should be used to compute homogeneity for.}

\item{cluster_cols}{Boolean, set to TRUE to perform column clustering.}

\item{fontsize}{Integer, base fontsize for the plot.}

\item{colors_annot}{Vector of colors to use.}
}
\value{
heatmap plot of top contributor genes for each informative mode of activation, ordered by decreasing contribution
}
\description{
Plots heatmap of top contributing genes for each mode
}
