% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_exploration.R
\name{plot_pathway_specificity}
\alias{plot_pathway_specificity}
\title{Plots barplot colored by cluster showing distribution of specificity between clusters for each activation mode}
\usage{
plot_pathway_specificity(
  PCA_object,
  module,
  meta,
  annot_name,
  colors_annot = NULL
)
}
\arguments{
\item{PCA_object}{List, output from run_activity_analysis().}

\item{module}{Character, module for which we want to plot activity.}

\item{meta}{Dataframe containing metadata, with cell labels as rownames.}

\item{annot_name}{Character, name of the column in meta that should be used to compute homogeneity for.}

\item{colors_annot}{Vector of colors to use}
}
\value{
UMAP plots (one by activation mode)
}
\description{
Plots barplot colored by cluster showing distribution of specificity between clusters for each activation mode
}
