% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_exploration.R
\name{plot_umap_annot}
\alias{plot_umap_annot}
\title{Plots UMAP colored by cell annotation}
\usage{
plot_umap_annot(umap, type = "Cell type", labels, title = NULL, colors = NULL)
}
\arguments{
\item{umap}{UMAP object}

\item{type}{Description of the labels that are plotted (cell type, unsupervised clusters,...)}

\item{labels}{Character vector, containing annotation for cells, in the same order as cells in the umap object}

\item{title}{Title of the plot.}

\item{colors}{List of colors, else default custom palette colors will be used}
}
\value{
UMAP plot
}
\description{
Plots UMAP colored by cell annotation
}
