% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_exploration.R
\name{plot_umap_gene}
\alias{plot_umap_gene}
\title{Plots UMAP colored by expression level of a gene}
\usage{
plot_umap_gene(umap, expr_mat, gene)
}
\arguments{
\item{umap}{UMAP object}

\item{expr_mat}{Numeric matrix, containing counts, normalized or not.}

\item{gene}{Character, gene for which we want to plot expression}
}
\value{
UMAP plot
}
\description{
Plots UMAP colored by expression level of a gene
}
