% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_gmt}
\alias{read_gmt}
\title{Read GMT file}
\usage{
read_gmt(gmt_file_name)
}
\arguments{
\item{gmt_file_name}{Character.}
}
\value{
List of modules, each associated with a character vector containing genes constituting the module.
}
\description{
Imports specified GMT file as a list
}
\examples{
read_gmt(system.file("extdata", "h.all.v7.4.symbols.gmt", package = "MAYA"))
}
