% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_umap}
\alias{run_umap}
\title{Run UMAP with fixed seed for reproducibility}
\usage{
run_umap(mat)
}
\arguments{
\item{mat}{Numeric matrix, output from compute_activity function.}
}
\value{
UMAP object
}
\description{
Run UMAP with fixed seed for reproducibility
}
