% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_0_1}
\alias{scale_0_1}
\title{Scale matrix rows between 0 and 1}
\usage{
scale_0_1(X)
}
\arguments{
\item{X}{Numeric matrix.}
}
\value{
Numeric matrix.
}
\description{
The minimum value of each row is set to 0 and the maximum to 1.
}
\examples{
raw_mat <- matrix(1:20, ncol = 5, nrow = 4)
norm_mat <- scale_0_1(raw_mat)
}
