% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{mayan_to_gregorian}
\alias{mayan_to_gregorian}
\title{Converts mayan long count date to gregorian date}
\usage{
mayan_to_gregorian(date = NULL, output = c("calendar", "named_vector",
  "astronomical", "number"))
}
\arguments{
\item{date}{Character vector}

\item{output}{Expected output}
}
\description{
Converts mayan long count date to gregorian date
}
\examples{

mayan_to_gregorian("0.0.0.0.0")
mayan_to_gregorian("13.0.0.0.0")

}
